+++ CHANGELOG.md
@@ -6,6 +6,7 @@
    `fmt_response_times()`, `fmt_percentiles()`, and `fmt_status_codes()`
  - display `GooseStats` with fmt::Display (ie `print!("{}", goose_stats);`)
  - make it possible to pass a closure to GooseTask::new
+ - fix display of `GooseError` and `GooseTaskError`
 
 ## 0.9.0 July 23, 2020
  - fix code documentation, requests are async and require await